pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools",
  "setuptools-scm",
]

[project]
name = "cachi2"
license = {text = "GPLv3+"}
requires-python = ">=3.9"
classifiers = [
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dynamic = [
  "version",
]
dependencies = [
  "aiohttp",
  "aiohttp-retry",
  "backoff",
  "beautifulsoup4",
  "gitpython",
  "gemlock_parser @ https://github.com/containerbuildsystem/gemlock-parser/archive/40b200f5e0b37a7b1e604fd3adb7d67c59f786d4.zip",
  "packageurl-python",
  "packaging",
  "pydantic",
  "pypi-simple",
  "pyyaml",
  "reflink",
  "requests",
  "semver",
  "setuptools",
  "tomli",
  "typer",
]
[project.optional-dependencies]
dev = [
  "bandit",
  "black",
  "click",
  "flake8",
  "flake8-docstrings",
  "isort[colors]",
  "mypy",
  "typing-extensions",
]
test = [
  "GitPython",
  "jsonschema",
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "pytest-env",
  "pyyaml",
]
[project.scripts]
cachi2 = "cachi2.interface.cli:app"

[tool.setuptools]
packages = ["cachi2"]

[tool.setuptools_scm]
version_scheme = "post-release"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
plugins = ["pydantic.mypy"]
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true

[[tool.mypy.overrides]]
module = ["semver"]
ignore_missing_imports = true

[tool.coverage.report]
skip_covered = true
show_missing = true
fail_under = 90
exclude_lines = [
  "def __repr__",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "return NotImplemented",
]
